requirements.txt
bcrypt==4.0.1
bidict==0.22.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
eventlet==0.33.3
Flask==1.1.2
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
greenlet==3.0.0
h11==0.14.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
python-engineio==4.8.0
python-socketio==5.10.0
simple-websocket==1.0.0
six==1.16.0
Werkzeug==0.16.1
wsproto==1.2.0
